+++ src/pages/[dao]/proposals.tsx
@@ -1,4 +1,4 @@
-import DAOLayout from "@/components/Layout/dao/layout";
+import DAOLayout from "@/components/Layout/dao";
 import { Container } from "@/components/common/BaseContainer";
 import { Box } from "@/components/common/Box";
 import { Button } from "@/components/common/Button";
